Refactors custom dialog library, thread managment, and 2FA QR dialog
- Moved `WalletThreadManager` class from `gui.py` to a seperate
`thread_manager.py` module.

- Moved `QRCodeUtils` from `interface_util.py` to a seperate
`qr_code_util.py` module,

- Replaced the old pygame 2FA QR dialog with a unified
`show_2FA_QR_dialog` method in the `Dialogs` class. This ensures that
it is consistent with the wallet client's GUI mode which uses our
custom tkinter dialog library.

- Added a `_2FA_QR_Dialog` class to `qr_code_util.py` which initializes
the 2FA QR Code dialog, along with it's associated variables and
references.

- Updated `handle_new_encrypted_wallet` function in `wallet_client.py`
to utilize the new `_2FA_QR_Dialog` class and manage the 2FA flow more
effectively.

- Added methods to the `DialogFunctions` Class to handle the 2FA dialog
interactions, including countdown timer, and reveal token
functionality. This also includes logic for a context menu and hotkeys,
which are used for selecting or copying the 2FA token.

- The `WalletThreadManager` class has been updated for non-GUI
contexts. It checks for a `root.after` event loop and, if not found, it
falls back to a dedicated polling thread.

- The `CustomDialog` class can now run standalone without a parent
window. It can now be instantiated without a parent window,
automatically creating a temporary root and centers itself on the
screen. This is usful for creating dialog windows when in CLI mode.